Move the `hooks-client` module under the `@bayou` namespace.
