+++ local-modules/doc-client/EditorComplex.js
@@ -8,7 +8,7 @@ import { Provider } from 'react-redux';
 
 import { SplitKey } from '@bayou/api-common';
 import { ClientStore } from '@bayou/data-model-client';
-import { Hooks } from 'hooks-client';
+import { Hooks } from '@bayou/hooks-client';
 import { Condition } from '@bayou/promise-util';
 import { BayouKeyHandlers, QuillProm } from '@bayou/quill-util';
 import { Logger } from '@bayou/see-all';
